python/django/mysite/requirements.txt
amqp==5.3.1
asgiref==3.8.1
aws-opentelemetry-distro==0.6.0
billiard==4.2.1
celery==5.3.6
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Django==5.1.7
googleapis-common-protos==1.65.0
grpcio==1.67.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.4.0
kombu==5.4.2
opentelemetry-api==1.27.0
opentelemetry-distro==0.48b0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-aio-pika==0.48b0
opentelemetry-instrumentation-aiohttp-client==0.48b0
opentelemetry-instrumentation-aiopg==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-asyncio==0.48b0
opentelemetry-instrumentation-asyncpg==0.48b0
opentelemetry-instrumentation-aws-lambda==0.48b0
opentelemetry-instrumentation-boto==0.48b0
opentelemetry-instrumentation-boto3sqs==0.48b0
opentelemetry-instrumentation-botocore==0.48b0
opentelemetry-instrumentation-cassandra==0.48b0
opentelemetry-instrumentation-celery==0.48b0
opentelemetry-instrumentation-confluent-kafka==0.48b0
opentelemetry-instrumentation-dbapi==0.48b0
opentelemetry-instrumentation-django==0.48b0
opentelemetry-instrumentation-elasticsearch==0.48b0
opentelemetry-instrumentation-falcon==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-grpc==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
opentelemetry-instrumentation-jinja2==0.48b0
opentelemetry-instrumentation-kafka-python==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-instrumentation-mysql==0.48b0
opentelemetry-instrumentation-mysqlclient==0.48b0
opentelemetry-instrumentation-pika==0.48b0
opentelemetry-instrumentation-psycopg2==0.48b0
opentelemetry-instrumentation-pymemcache==0.48b0
opentelemetry-instrumentation-pymongo==0.48b0
opentelemetry-instrumentation-pymysql==0.48b0
opentelemetry-instrumentation-pyramid==0.48b0
opentelemetry-instrumentation-redis==0.48b0
opentelemetry-instrumentation-remoulade==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-sqlite3==0.48b0
opentelemetry-instrumentation-starlette==0.48b0
opentelemetry-instrumentation-system-metrics==0.48b0
opentelemetry-instrumentation-threading==0.48b0
opentelemetry-instrumentation-tornado==0.48b0
opentelemetry-instrumentation-tortoiseorm==0.48b0
opentelemetry-instrumentation-urllib==0.48b0
opentelemetry-instrumentation-urllib3==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-propagator-b3==1.27.0
opentelemetry-propagator-jaeger==1.27.0
opentelemetry-propagator-ot-trace==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-sdk-extension-aws==2.0.2
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==24.1
prompt_toolkit==3.0.50
protobuf==4.25.5
psutil==6.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uWSGI==2.0.28
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
zipp==3.20.2
